+++ starkbank/taxpayment/__taxpayment.py
@@ -34,7 +34,7 @@ class TaxPayment(Resource):
         self.bar_code = bar_code
         self.description = description
         self.tags = tags
-        self.scheduled = check_datetime(scheduled)
+        self.scheduled = check_date(scheduled)
         self.status = status
         self.amount = amount
         self.fee = fee